+++ apps/web/resources/views/components/ui/button.blade.php
@@ -1,0 +1,15 @@
+@props([
+    'href' => null,
+    'type' => 'button',
+    'variant' => 'primary',
+])
+
+@php
+    $classes = 'teelle-button teelle-button--'.$variant;
+@endphp
+
+@if ($href)
+    <a href="{{ $href }}" {{ $attributes->class($classes) }}>{{ $slot }}</a>
+@else
+    <button type="{{ $type }}" {{ $attributes->class($classes) }}>{{ $slot }}</button>
+@endif